requirements.txt
# ComfyUI INSTARAW Dependencies
# Note: torch and numpy are already included with ComfyUI

# API nodes
requests
Pillow

# Mask processing (optional but recommended for GrowMaskWithBlur)
kornia
scipy
tqdm

#google-genai>=1.51.0
google-genai~=1.43.0
google-api-core

# Performance enhancement for model downloads from Hugging Face
huggingface_hub[hf_xet]

onnxruntime

piexif

scikit-image
lpips

pyexiftool

#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=mediapipe/opensource_only/requirements_lock_3_12.txt mediapipe/opensource_only/requirements.txt
#
absl-py==2.1.0
    # via -r mediapipe/opensource_only/requirements.txt
attrs==24.2.0
    # via -r mediapipe/opensource_only/requirements.txt
cffi==1.17.1
    # via sounddevice
contourpy==1.3.0
    # via matplotlib
cycler==0.12.1
    # via matplotlib
flatbuffers==24.3.25
    # via -r mediapipe/opensource_only/requirements.txt
fonttools==4.54.1
    # via matplotlib
jax==0.4.30
    # via -r mediapipe/opensource_only/requirements.txt
jaxlib==0.4.30
    # via
    #   -r mediapipe/opensource_only/requirements.txt
    #   jax
kiwisolver==1.4.7
    # via matplotlib
matplotlib==3.9.2
    # via -r mediapipe/opensource_only/requirements.txt
ml-dtypes==0.5.0
    # via
    #   jax
    #   jaxlib
numpy==1.26.4
    # via
    #   -r mediapipe/opensource_only/requirements.txt
    #   contourpy
    #   jax
    #   jaxlib
    #   matplotlib
    #   ml-dtypes
    #   opencv-contrib-python
    #   scipy
opencv-contrib-python==4.10.0.84
    # via -r mediapipe/opensource_only/requirements.txt
opt-einsum==3.4.0
    # via jax
packaging==24.1
    # via matplotlib
pillow==10.4.0
    # via matplotlib
protobuf==4.25.5
    # via -r mediapipe/opensource_only/requirements.txt
pycparser==2.22
    # via cffi
pyparsing==3.1.4
    # via matplotlib
python-dateutil==2.9.0.post0
    # via matplotlib
scipy==1.13.1
    # via
    #   jax
    #   jaxlib
sentencepiece==0.2.0
    # via -r mediapipe/opensource_only/requirements.txt
six==1.16.0
    # via python-dateutil
sounddevice==0.5.0
    # via -r mediapipe/opensource_only/requirements.txt

# MediaPipe - pinned to 0.10.14 for compatibility with comfyui_controlnet_aux
# Newer versions removed mp.solutions API which breaks face mesh preprocessor
mediapipe==0.10.14